Move types::Value into its own module.
